vecs_to_structured_reference_string and kzgsettings_to_structured_reference_string moved to srs.rs
