+++ src/srs.rs
@@ -1,5 +1,10 @@
+use crate::commitments::kzg::StructuredReferenceString;
 use crate::compress::{decompress_g1_point, decompress_g2_point};
-use crate::{blst_fp, blst_fp2, blst_p1, blst_p2, G2Point, KZGSettings, G1};
+use crate::math::{elliptic_curve::traits::FromAffine, traits::ByteConversion};
+use crate::{
+    blst_fp, blst_fp2, blst_p1, blst_p2, BLS12381FieldElement, BLS12381TwistCurveFieldElement,
+    G2Point, KZGSettings, G1, NUM_G1_POINTS, NUM_G2_POINTS,
+};
 use core::ptr::null_mut;
 use std::fs::File;
 use std::io::{self, BufRead, BufReader, Lines};
@@ -184,6 +189,89 @@ pub fn load_trusted_setup_file(path: &str) -> io::Result<KZGSettings> {
     // TODO - return the KZGSettings:
     // convert vector of g1 and g2 point to lambdaworks format
     Ok(settings)
+}
+
+pub fn vecs_to_structured_reference_string(
+    g1_points: &[G1],
+    g2_points: &[G2Point],
+) -> StructuredReferenceString<G1, G2Point> {
+    StructuredReferenceString::<G1, G2Point>::new(g1_points, g2_points)
+}
+
+pub fn kzgsettings_to_structured_reference_string(
+    s: &KZGSettings,
+) -> StructuredReferenceString<G1, G2Point> {
+    let g1_values = s.g1_values;
+    let g2_values = s.g2_values;
+
+    let g1_points_slice: [blst_p1; NUM_G1_POINTS] = unsafe { *g1_values.cast() };
+    let g2_points_slice: [blst_p2; NUM_G2_POINTS] = unsafe { *g2_values.cast() };
+
+    let g1_points: Vec<G1> = g1_points_slice
+        .iter()
+        .map(|point| {
+            let x_field = point
+                .x
+                .l
+                .iter()
+                .flat_map(|e| e.to_be_bytes())
+                .collect::<Vec<u8>>();
+            let x = BLS12381FieldElement::from_bytes_be(&x_field).unwrap();
+
+            let y_field = point
+                .y
+                .l
+                .iter()
+                .flat_map(|e| e.to_be_bytes())
+                .collect::<Vec<u8>>();
+            let y = BLS12381FieldElement::from_bytes_be(&y_field).unwrap();
+            G1::from_affine(x, y).unwrap()
+        })
+        .collect();
+
+    let g2_points: Vec<G2Point> = g2_points_slice
+        .iter()
+        .map(|point| {
+            let [x0, x1] = point.x.fp;
+            let [y0, y1] = point.y.fp;
+            //let z = point.z;
+
+            let x0_field = BLS12381FieldElement::from_bytes_be(
+                &x0.l
+                    .iter()
+                    .flat_map(|e| e.to_be_bytes())
+                    .collect::<Vec<u8>>(),
+            )
+            .unwrap();
+            let x1_field = BLS12381FieldElement::from_bytes_be(
+                &x1.l
+                    .iter()
+                    .flat_map(|e| e.to_be_bytes())
+                    .collect::<Vec<u8>>(),
+            )
+            .unwrap();
+            let y0_field = BLS12381FieldElement::from_bytes_be(
+                &y0.l
+                    .iter()
+                    .flat_map(|e| e.to_be_bytes())
+                    .collect::<Vec<u8>>(),
+            )
+            .unwrap();
+            let y1_field = BLS12381FieldElement::from_bytes_be(
+                &y1.l
+                    .iter()
+                    .flat_map(|e| e.to_be_bytes())
+                    .collect::<Vec<u8>>(),
+            )
+            .unwrap();
+
+            let x = BLS12381TwistCurveFieldElement::new([x0_field, x1_field]);
+            let y = BLS12381TwistCurveFieldElement::new([y0_field, y1_field]);
+            G2Point::from_affine(x, y).unwrap()
+        })
+        .collect();
+
+    StructuredReferenceString::<G1, G2Point>::new(&g1_points, &g2_points)
 }
 
 #[cfg(test)]
